Translate Model unit tests to TS. https://online.sbis.ru/opendoc.html?guid=119ab123-3c10-4c8e-83a2-da4c1c957dc4
